Create Task 2 Q1
This is the full code for Question 1 in Task 2, it can explain the process of how to solve the sudoku(e.g. put 1 in row 9 column 7). This program explains the process step by step using the --explain function. In-Order to make this program work. If you are using Spyder, please input --explain in the Command Line Options in General Setting, Configuration Per File in the Run tab