Create Task 3 test 1
In this version of the code, I have sucessfully use the back track solution to print out the final answer of the sudoku. However this doesn't meet the requirement of the Task. Where it requires us to use Wavefront propagation. Therefore the solving technique of this version needs to be modify to meet the requirement.